+++ core/creators/inline.js
@@ -15,6 +15,10 @@
 	 *		...
 	 *		CKEDITOR.inline( 'content' );
 	 *
+	 * It is also possible to create inline editor from textarea element. If you do so
+	 * an additional `div` with editable content will be created directly after the
+	 * `textarea` element and that `textarea` element will be hidden.
+	 *
 	 * @param {Object/String} element The DOM element or its ID.
 	 * @param {Object} [instanceConfig] The specific configurations to apply to this editor instance.
 	 * See {@link CKEDITOR.config}.
@@ -30,11 +34,30 @@
 		if ( element.getEditor() )
 			throw 'The editor instance "' + element.getEditor().name + '" is already attached to the provided element.';
 
-		var editor = new CKEDITOR.editor( instanceConfig, element, CKEDITOR.ELEMENT_MODE_INLINE );
+		var editor = new CKEDITOR.editor( instanceConfig, element, CKEDITOR.ELEMENT_MODE_INLINE ),
+			textarea = element.is( 'textarea' ) ? element : null;
 
-		// Initial editor data is simply loaded from the page element content to make
-		// data retrieval possible immediately after the editor creation.
-		editor.setData( element.getHtml(), null, true );
+		if ( textarea ) {
+			editor.setData( textarea.getValue(), null, true );
+
+			//Change element from textarea to div
+			element = CKEDITOR.dom.element.createFromHtml(
+				'<div contenteditable="' + !!editor.readOnly + '" class="cke_textarea_inline">' +
+					textarea.getValue() +
+				'</div>',
+				CKEDITOR.document );
+
+			element.insertAfter( textarea );
+			textarea.hide();
+
+			// Attaching the concrete form.
+			if ( textarea.$.form )
+				editor._attachToForm();
+		} else {
+			// Initial editor data is simply loaded from the page element content to make
+			// data retrieval possible immediately after the editor creation.
+			editor.setData( element.getHtml(), null, true );
+		}
 
 		// Once the editor is loaded, start the UI.
 		editor.on( 'loaded', function() {
@@ -53,6 +76,7 @@
 			editor.resetDirty();
 
 			editor.fire( 'contentDom' );
+
 			// Inline editing defaults to "wysiwyg" mode, so plugins don't
 			// need to make special handling for this "mode-less" environment.
 			editor.mode = 'wysiwyg';
@@ -68,9 +92,18 @@
 
 		// Handle editor destroying.
 		editor.on( 'destroy', function() {
+			// Remove container from DOM if inline-textarea editor.
+			// Show <textarea> back again.
+			if ( textarea ) {
+				editor.container.clearCustomData();
+				editor.container.remove();
+				textarea.show();
+			}
+
 			editor.element.clearCustomData();
+
 			delete editor.element;
-		});
+		} );
 
 		return editor;
 	};
@@ -94,7 +127,9 @@
 					// the instance settings and eventually cancel the creation.
 
 					data = {
-						element: el, config: {} };
+						element: el,
+						config: {}
+					};
 
 					if ( CKEDITOR.fire( 'inline', data ) !== false )
 						CKEDITOR.inline( el, data.config );
@@ -105,9 +140,8 @@
 
 	CKEDITOR.domReady( function() {
 		!CKEDITOR.disableAutoInline && CKEDITOR.inlineAll();
-	});
+	} );
 })();
-
 
 /**
  * Avoid creating editor automatically on element which has attribute